a machine-specific rounding error fix
git-svn-id: https://svn.salilab.org/imp/trunk@1446 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/saxs/include/RadialDistributionFunction.h
@@ -50,7 +50,7 @@ public:
   void show(std::ostream &out=std::cout) const;
 
   Float get_max_pr_distance(void) { return max_pr_distance_; }
-  Float set_max_pr_distance(double max_pr_distance) {
+  Float set_max_pr_distance(Float max_pr_distance) {
     max_pr_distance_ = max_pr_distance;
     distribution_.reserve(dist2index(max_pr_distance_));
     return max_pr_distance_;
